// lib/types.ts
import { CoreMessage } from 'ai'

export type Message = CoreMessage & {
  id: string
}

export interface Chat extends Record<string, any> {
  id: string
  title: string
  createdAt: Date
  // userId: string
  path?: string
  messages: Message[]
  sharePath?: string
}


export interface Session {
  user: {
    id: string
    email: string
  }
}